+++ internal/services/db/db.go
@@ -10,8 +10,8 @@ const FileDir = "./originaldata/*.csv"
 type Dau struct {
 	Sourse      string
 	PartnerName string
-	Date        uint64
-	Dau         uint64
+	Date        int
+	Dau         int
 }
 type Chans struct {
 	IncomingDau      chan Dau
